Move provider-independent code to Umbraco.StorageProviders project
