+++ src/Umbraco.StorageProviders/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System;
+
+[assembly:CLSCompliant(false)]
