+++ src/Umbraco.StorageProviders/Umbraco.StorageProviders.csproj
@@ -6,13 +6,9 @@
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="Azure.Storage.Blobs" Version="12.9.1" />
         <PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="5.0.3" PrivateAssets="all" />
         <PackageReference Include="Nerdbank.GitVersioning" Version="3.4.231" PrivateAssets="all" />
-        <PackageReference Include="SixLabors.ImageSharp.Web.Providers.Azure" Version="1.0.3" />
         <PackageReference Include="SonarAnalyzer.CSharp" Version="8.27.0.35380" PrivateAssets="all" />
-    </ItemGroup>
-    <ItemGroup>
-      <ProjectReference Include="..\Umbraco.StorageProviders\Umbraco.StorageProviders.csproj" />
+        <PackageReference Include="Umbraco.Cms.Web.Common" Version="9.0.0-rc002" />
     </ItemGroup>
 </Project>
