+++ src/Umbraco.StorageProviders/version.json
@@ -1,0 +1,7 @@
+{
+    "version": "0.1",
+    "nugetPackageVersion": {
+        "semVer": 2
+    },
+    "pathFilters": ["."]
+}
